added safety to refresh.pl; refactor and testing

+++ scripts/refresh.pl
@@ -44,8 +44,10 @@ use vars qw(
 	     $opt_v
 	     $opt_s
 	     $opt_c
+	     $opt_r
+	     $opt_x
 	  );
-getopts('hvsc');
+getopts('hvscrx');
 
 ## Embedded help through perldoc.
 if( $opt_h ){
@@ -54,17 +56,25 @@ if( $opt_h ){
 }
 
 ## Take care of arguments.
-my $be_chatty = 0;
 my $do_sessions = 0;
 my $do_caches = 0;
-if ( $opt_v ){ $be_chatty = 1; }
+my $do_scratch = 0;
+my $execute = 0;
 if ( $opt_s ){ $do_sessions = 1; }
 if ( $opt_c ){ $do_caches = 1; }
+if ( $opt_r ){ $do_scratch = 1; }
+if ( $opt_x ){ $execute = 1; }
 
 ## Nothing at all? Then do everything.
-if ( ! $opt_s && ! $opt_c ){
+if ( ! $opt_s && ! $opt_c && ! $opt_r ){
+  ll("Will perform all tasks.");
   $do_sessions = 1;
   $do_caches = 1;
+  $do_scratch = 1;
+}
+
+if( $execute ){
+  ll("Execute flag set--will recursively delete filesystem items!");
 }
 
 ###
@@ -74,59 +84,17 @@ if ( ! $opt_s && ! $opt_c ){
 
 if( $do_sessions ){
 
-  ll("Starting session cleaning, please wait...") if $be_chatty;
-
-  ## Function to decide if the file or diectory has reached its
-  ## lifespan limit. Wired to one day (see variable below).
-  sub lifespan {
-
-    my $file = shift || die "need a file argument...";
-    my $retval = 0;
-
-    ## One day in seconds
-    #my $lifespan_limit_one_day = 1; # for debug
-    my $lifespan_limit_one_day = 86400;
-
-    ## Calc. time.
-    my $now = time;
-    my $mtime = stat($file)->mtime;
-    my $tdiff = $now - $mtime;
-
-    if( $tdiff >= $lifespan_limit_one_day ){
-      $retval = 1;
-    }
-
-    return $retval;
-  }
-
-  # ## Scrub out the generated images, maps, and directories (?!).
-  # sub temp_images{
-  #   if( (/\.png$/ || /\.gif$/ || /\.map$/ ) && lifespan($File::Find::name) ){
-  #     unlink $File::Find::name;
-  #   }elsif( /amigo/ && -d && lifespan($File::Find::name) ){
-  #     #remove_tree($File::Find::name, {safe => 1});
-  #     #rmtree($File::Find::name, {safe => 1});
-  #     rmtree($File::Find::name);
-  #   }
-  # }
-  # finddepth(\&temp_images, ($ENV{AMIGO_TEMP_IMAGE_DIR}));
-  # ll("Finished removing generated images.");
+  ll("Starting session cleaning...");
 
   ## Scrub out new session type files.
+  ll(" Removing session files.");
   sub new_sessions{
     if( /^cgisess_/ && lifespan($File::Find::name) ){
-      #ll(" unlink: " . $File::Find::name);
-      unlink $File::Find::name;
+      ll("  Removing: " . $File::Find::name);
+      unlink $File::Find::name if $execute;
     }
   }
-  finddepth(\&new_sessions, ($ENV{AMIGO_SESSIONS_ROOT_DIR}));
-  ll("Finished removing new session files.");
-
-  # ## Scrub out anything hanging out in scratch.
-  # # rmtree($ENV{AMIGO_SCRATCH_DIR},
-  # # 	 #{safe => 1, keep_root => 1} );
-  # # 	 {keep_root => 1} );
-  ll("Finished cleaning scratch directory: " . $ENV{AMIGO_SCRATCH_DIR});
+  finddepth(\&new_sessions, ($core->amigo_env('AMIGO_SESSIONS_ROOT_DIR')));
 
   ll("Finished session cleaning.");
 }
@@ -134,37 +102,68 @@ if( $do_sessions ){
 
 if( $do_caches ){
 
-  ll("Starting cache cleaning, please wait...");
-
-  my $core = AmiGO->new();
+  ll("Starting cache cleaning...");
 
   ## Clean out old KVStore caches.
-  $core->kvetch("Removing old KVStore caches...");
+  ll(" Removing KVStore caches...");
   foreach my $ca (@{ AmiGO::KVStore::list() }){
-    $core->kvetch("Removing: $ca");
-    #unlink $ca if -f $ca;
+    ll("  Removing: $ca");
+    unlink $ca if -f $ca && $execute;
   }
 
   ## Clean out old KVStore::Filesystem caches (not (currently) a
   ## subclass of KVStore, so not on its list).
-  $core->kvetch("Removing old KVStore::Filesystem caches...");
+  ll(" Removing KVStore::Filesystem caches...");
   foreach my $ca (@{ AmiGO::KVStore::Filesystem::list() }){
-    $core->kvetch("Removing: $ca");
-    #rmtree($ca, {safe => 1});
+    ll("  Removing tree: $ca");
+    rmtree($ca, {safe => 1}) if -d $ca && $execute;
   }
 
   ll("Finished cache cleaning.");
+}
+
+if( $do_scratch ){
+
+  ## Scrub out anything hanging out in scratch.
+  ll("Resetting scratch directory: " . $core->amigo_env('AMIGO_SCRATCH_DIR'));
+  rmtree($core->amigo_env('AMIGO_SCRATCH_DIR'),
+  	 #{safe => 1, keep_root => 1} );
+  	 {keep_root => 1} ) if $execute;
+
 }
 
 ###
 ### Subs.
 ###
 
+## Function to decide if the file or diectory has reached its
+## lifespan limit. Wired to one day (see variable below).
+sub lifespan {
+
+  my $file = shift || die "need a file argument...";
+  my $retval = 0;
+
+  ## One day in seconds
+  #my $lifespan_limit_one_day = 1; # for debug
+  my $lifespan_limit_one_day = 86400;
+
+  ## Calc. time.
+  my $now = time;
+  my $mtime = stat($file)->mtime;
+  my $tdiff = $now - $mtime;
+
+  if( $tdiff >= $lifespan_limit_one_day ){
+    $retval = 1;
+  }
+
+  return $retval;
+}
+
 ## Just a little printin' when feeling verbose.
 sub ll {
   my $str = shift || '';
-  print $str . "\n" if $be_chatty;
-  #$core->kvetch($str);
+  print $str . "\n" if $opt_v;
+  $core->kvetch($str);
 }
 
 
@@ -174,7 +173,7 @@ refresh.pl
 
 =head1 SYNOPSIS
 
-refresh.pl [-h] [-v] [-s] [-c]
+refresh.pl [-h] [-v] [-s] [-c] [-r] [-x]
 
 =head1 DESCRIPTION
 
@@ -197,6 +196,15 @@ and files that appear to be more than a day old.
 Perform routine cleaning of caches. This will remove any cached files
 from the server so that a new one will have to be fetched or
 generated.
+
+=item -r
+
+Reset the scratch directory.
+
+=item -x
+
+Since we're changing the filesystem and doing recursive deletes and
+the like, we won't actually do anything unless we have this flag.
 
 =back
